refactor(fleet): split policy form parsing out of the policy page module
fleet_policy_page.py crossed the 2000-line module ceiling (2088). Move
Submission, parse_form, build_document, and the scoped-edit helpers into
fleet_policy_form.py and re-export them from the page so callers and tests
keep their import paths. No behavior change.
